dependencies: move to dependdencies package
