+++ dependencies/src/main/kotlin/dependencies/Testing.kt
@@ -1,4 +1,4 @@
-package com.louiscad.splitties
+package dependencies
 
 object Testing {
     private const val strikt_version = "+"
