+++ refreshVersions/versions.properties
@@ -8,7 +8,7 @@ plugin.org.gradle.kotlin.kotlin-dsl=1.2.9
 #                       # available=1.3.1
 plugin.com.gradle.plugin-publish=0.10.0
 #                    # available=0.10.1
-plugin.de.fayard.refreshVersions=0.8.1
+plugin.de.fayard.refreshVersions=0.8.4
 plugin.com.gradle.build-scan=2.4.1
 #                # available=3.0
 version.org.jetbrains.kotlin=1.3.41
